+++ Sources/Utils/UserDefaultsManager/DefaultsKeys+UserDefaults.swift
@@ -17,11 +17,11 @@ extension DefaultsKeys {
 extension DefaultsKeys {
     var agreePolicy: DefaultsKey<Bool> { .init("agreePolicy", defaultValue: false) }
 
-    var splashLoadTime: DefaultsKey<Int> { .init("splashLoadTime", defaultValue: 0) }
-    var splashPullInterval: DefaultsKey<Int> { .init("splashPullInterval", defaultValue: 0) }
+    var splashLoadTime: DefaultsKey<Double> { .init("splashLoadTime", defaultValue: 0) }
+    var splashPullInterval: DefaultsKey<Double> { .init("splashPullInterval", defaultValue: 0) }
 
-    var adShowTime: DefaultsKey<Int> { .init("adShowTime", defaultValue: 0) }
-    var adLoadTime: DefaultsKey<Int> { .init("adLoadTime", defaultValue: 0) }
-    var adPullInterval: DefaultsKey<Int> { .init("adPullInterval", defaultValue: 0) }
-    var adMinInterval: DefaultsKey<Int> { .init("adMinInterval", defaultValue: 0) }
+    var adShowTime: DefaultsKey<Double> { .init("adShowTime", defaultValue: 0) }
+    var adLoadTime: DefaultsKey<Double> { .init("adLoadTime", defaultValue: 0) }
+    var adPullInterval: DefaultsKey<Double> { .init("adPullInterval", defaultValue: 0) }
+    var adMinInterval: DefaultsKey<Double> { .init("adMinInterval", defaultValue: 0) }
 }
